+++ src/main/java/com/midproject/schoolregistrationsystem/Repositories/CourseRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CourseRepository extends JpaRepository<Course, Integer>
-{
+public interface CourseRepository extends JpaRepository<Course, Long>{
+
 
 }
